moved all tests to their own files. refactored setup and teardown for all test suites
