+++ src/test/basic.test.ts
@@ -9,20 +9,14 @@ import {
   extSuffix,
   AffectedSettings
 } from '../models';
-import {
-  setupTestSuite,
-  teardownTestSuite
-} from './lib/setup-teardown-test-suite';
+import { allSetupAndTeardown } from './lib/setup-teardown-test-suite';
+import { getExtension } from './lib/helpers';
 
 suite('Basic Extension Tests', () => {
-  let extension: vscode.Extension<any>;
+  let extension = <vscode.Extension<any>>getExtension();
   let originalValues = <IPeacockSettings>{};
 
-  suiteSetup(async () => {
-    extension = await setupTestSuite(extension, originalValues);
-  });
-
-  suiteTeardown(() => teardownTestSuite(originalValues));
+  allSetupAndTeardown(originalValues);
 
   test('Extension loads in VSCode and is active', done => {
     // Hopefully a 200ms timeout will allow the extension to activate within Windows
