+++ src/test/lib/helpers.ts
@@ -1,21 +1,9 @@
 import * as vscode from 'vscode';
 import * as sinon from 'sinon';
-import {
-  Commands,
-  ColorSettings,
-  Sections,
-  IPeacockElementAdjustments
-} from '../../models';
+import { Commands, ColorSettings, Sections } from '../../models';
 
-export const executeCommand = vscode.commands.executeCommand;
-
-export const noopElementAdjustments = <IPeacockElementAdjustments>{
-  activityBar: 'none',
-  statusBar: 'none',
-  titleBar: 'none'
-};
-
-export function getExtension(extension: vscode.Extension<any>) {
+export function getExtension() {
+  let extension: vscode.Extension<any> | undefined;
   const ext = vscode.extensions.getExtension('johnpapa.vscode-peacock');
   if (!ext) {
     throw new Error('Extension was not found.');
